began refactoring yolov8.py
